+++ src/components/molecules/alert/index.js
@@ -40,7 +40,7 @@ var fadeOut = (0, _styledComponents.keyframes)(["0%{opacity:1;}100%{opacity:0;}"
 var AlertWrapper = _styledComponents.default.div.withConfig({
   displayName: "alert__AlertWrapper",
   componentId: "sc-1nkifwv-0"
-})(["", ";margin:", ";border-radius:", ";overflow:hidden;background-color:", ";color:", ";display:flex;[href]{color:", ";}.k-Alert__text{padding:", " ", ";flex:1 0 0;font-size:", ";display:flex;align-items:center;}.k-Alert__iconBadge{margin-right:", ";background-color:", ";border-color:", ";min-width:", ";min-height:", ";border-radius:", ";@media (min-width:", "){margin-right:", ";}}button{display:flex;flex:0 0 auto;transition:all 0.2s ease;align-self:stretch;align-items:center;padding-right:", ";@media (min-width:", "){padding-right:", ";}svg,svg path{transition:fill 0.2s ease;fill:", ";:hover{fill:", ";}}&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}a{", ";color:inherit;text-decoration:underline;}&.k-Alert--success{background-color:", ";[href]{color:", ";}.k-Alert__iconBadge{background-color:", ";border-color:", ";}button{svg,svg path{fill:", ";:hover{fill:", ";}}}}&.k-Alert--error{background-color:", ";[href]{color:", ";}.k-Alert__iconBadge{background-color:", ";border-color:", ";}button{svg,svg path{fill:", ";:hover{fill:", ";}}}}&.k-Alert--warning{background-color:", ";[href]{color:", ";}.k-Alert__iconBadge{background-color:", ";border-color:", ";}button{svg,svg path{fill:", ";:hover{fill:'#A47600';}}}}&.k-Alert--shouldHide{pointer-events:none;animation:", " 0.4s cubic-bezier(0.895,0.03,0.685,0.22) forwards;}"], _typographyConfig.default.fontStyles.light, (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(8), _colorsConfig.default.primary5, _colorsConfig.default.font1, _colorsConfig.default.primary1, (0, _typography.pxToRem)(18), (0, _typography.pxToRem)(20), (0, _typography.stepToRem)(-1), (0, _typography.pxToRem)(10), _colorsConfig.default.primary1, _colorsConfig.default.primary4, (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(_screenConfig.ScreenConfig.S.min), (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(_screenConfig.ScreenConfig.S.min), (0, _typography.pxToRem)(20), _colorsConfig.default.primary1, _colorsConfig.default.primary3, _colorsConfig.default.background1, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(-4), _colorsConfig.default.primary4, _typographyConfig.default.fontStyles.bold, _colorsConfig.default.tertiary1, _colorsConfig.default.valid, _colorsConfig.default.valid, _colorsConfig.default.tertiary2, _colorsConfig.default.valid, _colorsConfig.default.tertiary2, _colorsConfig.default.error2, _colorsConfig.default.error, _colorsConfig.default.error, _colorsConfig.default.error3, _colorsConfig.default.error, _colorsConfig.default.error4, _colorsConfig.default.warning2, _colorsConfig.default.orange3, _colorsConfig.default.orange3, _colorsConfig.default.orange, _colorsConfig.default.warning, fadeOut);
+})(["--alert-gap:", ";@media ", "{--alert-gap:", ";}", ";margin:", ";border-radius:", ";overflow:hidden;background-color:", ";color:", ";display:flex;align-items:center;gap:var(--alert-gap);padding:", " var(--alert-gap);&.k-Alert--hasCloseButton{--alert-close-width:calc(var(--alert-gap) + ", ");}&.k-Alert--hasIcon{--alert-icon-width:calc(var(--alert-gap) + ", ");}&.k-Alert--center{&.k-Alert--hasIcon:not(.k-Alert--hasCloseButton) .k-Alert__text{padding-right:var(--alert-icon-width);}&.k-Alert--hasCloseButton:not(.k-Alert--hasIcon) .k-Alert__text{padding-left:var(--alert-close-width);}&.k-Alert--hasCloseButton.k-Alert--hasIcon .k-Alert__text{padding-right:calc(var(--alert-icon-width) - var(--alert-close-width));}.k-Alert__text{text-align:center;flex-basis:auto;}}.k-Alert__text{flex:1 0 calc(100% - var(--alert-close-width,0px) - var(--alert-icon-width,0px));font-size:", ";}.k-Alert__icon{.k-Alert__iconBadge{background-color:", ";border-color:", ";min-width:", ";min-height:", ";border-radius:", ";}}.k-Alert__closeButton{flex:0 0 ", ";transition:all 0.2s ease;display:flex;align-self:stretch;align-items:center;margin:", " calc(var(--alert-gap) * -1);padding:", " var(--alert-gap);svg,svg path{transition:fill 0.2s ease;fill:", ";:hover{fill:", ";}}&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}a{", ";color:", ";text-decoration:underline;}&.k-Alert--success{background-color:", ";[href]{color:", ";}.k-Alert__iconBadge{background-color:", ";border-color:", ";}.k-Alert__closeButton{svg,svg path{fill:", ";:hover{fill:", ";}}}}&.k-Alert--error{background-color:", ";[href]{color:", ";}.k-Alert__iconBadge{background-color:", ";border-color:", ";}.k-Alert__closeButton{svg,svg path{fill:", ";:hover{fill:", ";}}}}&.k-Alert--warning{background-color:", ";[href]{color:", ";}.k-Alert__iconBadge{background-color:", ";border-color:", ";}.k-Alert__closeButton{svg,svg path{fill:", ";:hover{fill:'#A47600';}}}}&.k-Alert--shouldHide{pointer-events:none;animation:", " 0.4s cubic-bezier(0.895,0.03,0.685,0.22) forwards;}"], (0, _typography.pxToRem)(10), _screenConfig.mq.tabletAndDesktop, (0, _typography.pxToRem)(20), _typographyConfig.default.fontStyles.light, (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(8), _colorsConfig.default.primary5, _colorsConfig.default.font1, (0, _typography.pxToRem)(18), (0, _typography.pxToRem)(17), (0, _typography.pxToRem)(30), (0, _typography.stepToRem)(-1), _colorsConfig.default.primary1, _colorsConfig.default.primary4, (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(24), (0, _typography.pxToRem)(17), (0, _typography.pxToRem)(-18), (0, _typography.pxToRem)(18), _colorsConfig.default.primary1, _colorsConfig.default.primary3, _colorsConfig.default.background1, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(-2), _colorsConfig.default.primary4, _typographyConfig.default.fontStyles.bold, _colorsConfig.default.primary1, _colorsConfig.default.tertiary1, _colorsConfig.default.valid, _colorsConfig.default.valid, _colorsConfig.default.tertiary2, _colorsConfig.default.valid, _colorsConfig.default.tertiary2, _colorsConfig.default.error2, _colorsConfig.default.error, _colorsConfig.default.error, _colorsConfig.default.error3, _colorsConfig.default.error, _colorsConfig.default.error4, _colorsConfig.default.warning2, _colorsConfig.default.orange3, _colorsConfig.default.orange3, _colorsConfig.default.orange, _colorsConfig.default.warning, fadeOut);
 
 var Alert = function Alert(_ref) {
   var className = _ref.className,
@@ -54,7 +54,8 @@ var Alert = function Alert(_ref) {
       onAfterClose = _ref.onAfterClose,
       icon = _ref.icon,
       iconBadgeBorderColor = _ref.iconBadgeBorderColor,
-      others = (0, _objectWithoutProperties2.default)(_ref, ["className", "show", "error", "success", "warning", "closeButton", "closeButtonLabel", "children", "onAfterClose", "icon", "iconBadgeBorderColor"]);
+      center = _ref.center,
+      others = (0, _objectWithoutProperties2.default)(_ref, ["className", "show", "error", "success", "warning", "closeButton", "closeButtonLabel", "children", "onAfterClose", "icon", "iconBadgeBorderColor", "center"]);
 
   var _useState = (0, _react.useState)(false),
       _useState2 = (0, _slicedToArray2.default)(_useState, 2),
@@ -86,15 +87,17 @@ var Alert = function Alert(_ref) {
     ref: alertRef,
     role: "alert",
     className: (0, _classnames.default)('k-Alert', className, {
+      'k-Alert--center': center,
       'k-Alert--success': success,
       'k-Alert--error': error,
       'k-Alert--warning': warning,
-      'k-Alert--hasCloseButton': closeButton,
+      'k-Alert--hasCloseButton': !!closeButton,
+      'k-Alert--hasIcon': !!icon,
       'k-Alert--shouldHide': !isMounted
     })
-  }, others), /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, /*#__PURE__*/_react.default.createElement("div", {
-    className: "k-Alert__text"
-  }, icon && /*#__PURE__*/_react.default.createElement(_iconBadge.IconBadge, {
+  }, others), !!icon && /*#__PURE__*/_react.default.createElement("div", {
+    className: "k-Alert__icon"
+  }, /*#__PURE__*/_react.default.createElement(_iconBadge.IconBadge, {
     className: "k-Alert__iconBadge",
     children: icon,
     border: {
@@ -102,14 +105,16 @@ var Alert = function Alert(_ref) {
       color: iconBadgeBorderColor,
       style: 'solid'
     }
-  }), children), closeButton && /*#__PURE__*/_react.default.createElement("button", {
-    className: "k-u-reset-button"
+  })), /*#__PURE__*/_react.default.createElement("div", {
+    className: "k-Alert__text"
+  }, children), closeButton && /*#__PURE__*/_react.default.createElement("button", {
+    className: "k-Alert__closeButton k-u-reset-button"
   }, /*#__PURE__*/_react.default.createElement(_crossIconNext.CrossIconNext, {
     onClick: function onClick() {
       return setMounted(false);
     },
     title: closeButtonLabel
-  }))));
+  })));
 };
 
 exports.Alert = Alert;
@@ -122,7 +127,8 @@ Alert.propTypes = {
   closeButtonLabel: _propTypes.default.string,
   onAfterClose: _propTypes.default.func,
   icon: _propTypes.default.node,
-  iconBadgeBorderColor: _propTypes.default.string
+  iconBadgeBorderColor: _propTypes.default.string,
+  center: _propTypes.default.bool
 };
 Alert.defaultProps = {
   show: true,
@@ -132,6 +138,7 @@ Alert.defaultProps = {
   closeButton: false,
   closeButtonLabel: 'Close',
   onAfterClose: function onAfterClose() {},
-  icon: '',
-  iconBadgeBorderColor: _colorsConfig.default.primary4
+  icon: null,
+  iconBadgeBorderColor: _colorsConfig.default.primary4,
+  center: false
 };